+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.17.0] - 2021-04-02
 
 ### Changed
 
@@ -250,7 +250,8 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 [@yannham]: https://github.com/yannham
 [@zueve]: https://github.com/zueve
 
-[Unreleased]: https://github.com/theskumar/python-dotenv/compare/v0.16.0...HEAD
+[Unreleased]: https://github.com/theskumar/python-dotenv/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/theskumar/python-dotenv/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/theskumar/python-dotenv/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/theskumar/python-dotenv/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/theskumar/python-dotenv/compare/v0.13.0...v0.14.0